MINIFICPP-1151 fix easy 1stparty warnings
This closes #754

Remaining:
 - MSVC: multiple warning LNK4221: This object file does not define any previously undefined public symbols, so it will not be used by any link operation that consumes this library
 - MSVC/expression language: multiple warning C4566: character represented by universal-character-name '\u????' cannot be represented in the current code page (1252)
 - clang/expression language yacc: warning: fix-its can be applied.  Rerun with option '--update'. [-Wother]

Squashed commit of the following:

commit 9c555035ef7a3a3e8e9d12c00ea39b26f8f2d669
Author: Marton Szasz <szaszm01@gmail.com>
Date:   Wed Apr 15 18:30:10 2020 +0200

    MINIFICPP-1151 revert ListSFTP as its issue is already fixed on master

commit bacdadbafe4aca3fe320569c984d358c5712d014
Author: Marton Szasz <szaszm01@gmail.com>
Date:   Wed Apr 15 15:46:29 2020 +0200

    MINIFICPP-1151 fix invalid fread usage

commit 1b5ef640bbfb5421ee5e67e8c1a9eebb1aa2300e
Author: Marton Szasz <szaszm01@gmail.com>
Date:   Tue Apr 14 15:00:54 2020 +0200

    MINIFICPP-1151 windows (except expression language)

commit 96bfa5f2d8d8d419a0d677cf72b2662c0fb98160
Author: Marton Szasz <szaszm01@gmail.com>
Date:   Tue Apr 7 10:18:26 2020 +0200

    MINIFICPP-1151 linux, clang 10

    note: rocksdb compilation correctly fails with clang 10. We will need to
    upgrade or patch it.

commit aff657171ac197b85df6f5beea62d83a880622a1
Author: Marton Szasz <szaszm01@gmail.com>
Date:   Mon Feb 10 21:35:29 2020 +0100

    verified on clang6/linux

    There are some warnings due to overriding a ROCKSDB_ASSERT macro to
    throw and it's expanded inside `noexcept` functions. No idea how to fix
    that.

commit 265e3a796ba7384d2ed7a463ca18bbfa704ef654
Author: Marton Szasz <szaszm01@gmail.com>
Date:   Mon Feb 10 18:05:55 2020 +0100

    fix compilation errors

commit 5dcf968a07bf4d1d2d36cc90fbee39f16e02dea1
Author: Marton Szasz <szaszm01@gmail.com>
Date:   Mon Feb 10 15:58:05 2020 +0100

    MINIFICPP-1151 round #1: GCC 7 on ubuntu, default config

MINIFICPP-1151 fix ProcessGroup lifetime by making it temporary

+++ libminifi/include/core/state/Value.h
@@ -159,7 +159,7 @@ class UInt32Value : public Value {
   }
 
   virtual bool getValue(int &ref) {
-    if (value <= (std::numeric_limits<int>::max)()) {
+    if (value <= static_cast<uint32_t>((std::numeric_limits<int>::max)())) {
       ref = value;
       return true;
     }
@@ -323,7 +323,7 @@ class UInt64Value : public Value {
   }
 
   virtual bool getValue(int64_t &ref) {
-    if (value <= (std::numeric_limits<int64_t>::max)()) {
+    if (value <= static_cast<uint32_t>((std::numeric_limits<int64_t>::max)())) {
       ref = value;
       return true;
     }
